userRouter code from app to userRouter.js
