+++ 4-natours/starter/app.js
@@ -1,51 +1,15 @@
 const express = require('express');
 const morgan = require('morgan');
-const tourRouter = require('./routes/tourRoutes');
+const tourRouter = require('./routes/tourRouter');
+const userRouter = require('./routes/userRouter');
 
 const app = express();
 
 app.use(morgan('dev'));
 app.use(express.json());
 
-const getAllUsers = (req, res) => {
-  res.status(500).json({
-    status: 'error',
-    message: 'This route is not yet defined',
-  });
-};
-const createUser = (req, res) => {
-  res.status(500).json({
-    status: 'error',
-    message: 'This route is not yet defined',
-  });
-};
-const getUser = (req, res) => {
-  res.status(500).json({
-    status: 'error',
-    message: 'This route is not yet defined',
-  });
-};
-const updateUser = (req, res) => {
-  res.status(500).json({
-    status: 'error',
-    message: 'This route is not yet defined',
-  });
-};
-const deleteUser = (req, res) => {
-  res.status(500).json({
-    status: 'error',
-    message: 'This route is not yet defined',
-  });
-};
-
-const userRouter = express.Router();
-
 app.use('/api/v1/tours', tourRouter);
 app.use('/api/v1/users', userRouter);
-
-userRouter.route('/').get(getAllUsers).post(createUser);
-
-userRouter.route('/:id').get(getUser).patch(updateUser).delete(deleteUser);
 
 const port = 3000;
 app.listen(port, () => {
